backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
aiosmtplib>=3.0.0
python-multipart>=0.0.6
apscheduler>=3.10.4
google-api-python-client>=2.100.0
beautifulsoup4>=4.12.0
requests>=2.31.0
